requirements.txt
streamlit
pandas
supabase
streamlit-cookies-manager
google-cloud-vision
streamlit-geolocation
requests